--- server.ini
@@ -1,3 +1,0 @@
-[server]
-root=.
-static=static
