--- templates/index.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-
-        <!-- Latest compiled and minified CSS -->
-        <link   rel="stylesheet" 
-                href="//netdna.bootstrapcdn.com/bootstrap/3.1.0/css/bootstrap.min.css">
-
-        <!-- Optional theme -->
-        <link   rel="stylesheet" 
-                href="//netdna.bootstrapcdn.com/bootstrap/3.1.0/css/bootstrap-theme.min.css">
-        
-        <title>Seymour audio collections</title>
-
-        <style type="text/css">
-
-            .track {
-                margin:         2px;
-                padding:        4px;
-                text-align:     justify;
-            }
-        </style>
-    </head>
-
-    <body>
-    <!-- Nav header with a drop-down full of collections -->
-
-        <nav class="navbar navbar-default" role="navigation">
-            <div class="container-fluid">
-                <div class="navbar-header">
-                    <a class="navbar-brand" href="#">Seymour</a>
-                </div>
-                <div class="collapse navbar-collapse">
-                    <ul class="nav navbar-nav">
-                        <li class="dropdown">
-                            <a  href="#" 
-                                class="dropdown-toggle" 
-                                data-toggle="dropdown">
-                                <span id="collections-count"></span> collections<b class="caret"></b></a>
-                            <ul class="dropdown-menu" id="collections-menu">
-                            </ul>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </nav>
-
-        <div class="container container-fluid col-md-12">
-            <div class="row">
-                <div class="panel panel-default">
-                    <div class="panel-heading">
-                        <input type="hidden" id="offset" />
-                        <input type="hidden" id="limit" />
-                        <input type="hidden" id="track_count" />
-                        <input type="hidden" id="collection_id" value="{{ collection_id }}" />
-                        <ul class="pager">
-                            <li class="previous"><a href="#">&larr; Previous</a></li>
-                            <li class="form-group form-inline" id='track-form'> 
-                                <label for="track-min" id="collection-name" class="form-inline"></label> 
-                                <input id="track-min" type="text" class="form-control" style="width: 6em;"> 
-                                <label for="track-max"> to </label> 
-                                <span id="track-max" type="text" class="form-control-static" style="width: 6em;"></span>
-                                of <span id="track-total"></span>
-                            </li>
-
-                            <li class="next"><a href="#">Next &rarr;</a></li>
-                        </ul>
-                    </div>
-                    <div class="panel-body container">
-                    <div id="tracklist" class="list-group list-inline" ></div>
-
-                    </div>
-                </div>
-            </div>
-        </div>
-    </body>
-
-    <script src="//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js"></script>
-    
-    <!-- Latest compiled and minified JavaScript -->
-    <script src="//netdna.bootstrapcdn.com/bootstrap/3.1.0/js/bootstrap.min.js"></script>
-
-    <script src="/static/js/index.js"></script>
-</html>
